增加deposit第二版

+++ src/main/java/com/yupi/springbootinit/model/dto/deposit/DepositInfoAddRequest.java
@@ -32,7 +32,7 @@ public class DepositInfoAddRequest implements Serializable {
     /**
      *
      */
-    private Integer cardType;
+    private Long cardType;
 
     /**
      *
